bug fixed

+++ install.sh
@@ -7,26 +7,25 @@ bash_file=~/.bashrc
 
 # Checking if package is installed else install it
 function checkOrInstall(){
-	REQUIRED_PKG=$1
-    PKG_OK=$(dpkg-query -W --showformat='${Status}\n' $REQUIRED_PKG|grep "Already Installed")
+    PKG_OK=$(dpkg-query -W --showformat='${Status}\n' $1)
 
-	echo Checking for $REQUIRED_PKG: $PKG_OK
-
-	if [ "" = "$PKG_OK" ]; then
-		echo "No $REQUIRED_PKG. Setting up $REQUIRED_PKG."
-		sudo apt install -y $REQUIRED_PKG
+	if [ "install ok installed" != "$PKG_OK" ]; then
+		echo "No $1. Setting up $1."
+		sudo apt install -y $1
 	fi
 }
 
 # Find dir of python3 packages
 function searchPythonPackages(){
 	python_version=$(python3 --version)
-	packages=($(find / ! \( -path /mnt -prune \) -type $2 2> /dev/null | grep $1))
+	declare -a packages=($(find / ! \( -path /mnt -prune \) -type $2 2> /dev/null | grep $1))
+
+    echo ${packages[@]}
 	
 	for package in ${packages[@]}
 	do
 		if [[ "$package" == *"${python_version:7}"* ]]; then
-			echo $package
+			echo "$package"
 		fi
     done
 }
@@ -43,10 +42,16 @@ checkOrInstall python3-pip
 checkOrInstall grep
 
 # neovim
-wget https://github.com/neovim/neovim/releases/download/v0.7.2/nvim-linux64.deb
-sudo apt install -y ./nvim-linux64.deb
 
-rm -r nvim-linux64.deb
+NVIM_PKG_OK=$(dpkg-query -W --showformat='${Status}\n' nvim)
+
+if [ "install ok installed" != "$NVIM_PKG_OK" ]; then
+	echo "No $1. Setting up $1."
+	wget https://github.com/neovim/neovim/releases/download/v0.7.2/nvim-linux64.deb
+	sudo apt install -y ./nvim-linux64.deb
+
+	rm -r nvim-linux64.deb
+fi
 
 # powerline
 pip3 install --user powerline-status
